[Bronze I] Title: 운동회, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/25425. 운동회/운동회.cc
@@ -6,17 +6,17 @@ int main(void) {
     ios_base::sync_with_stdio(false);
     cin.tie(nullptr);
 
-    int N, M, a, K;
+    long long N, M, a, K;
     cin >> N >> M >> a >> K;
-    int num = a - K;
-    int mn, mx;
+    long long num = a - K;
+    long long mn, mx;
     if (num >= N) {
         mn = N;
     } else {
         mn = num + 1;
     }
 
-    int tmp = num / M;
+    long long tmp = num / M;
     tmp += (num % M) ? 1 : 0;
     mx = tmp + 1;
     cout << mn << ' ' << mx << '\n';
